Agentic AI Q4/Class Code/Class-4/pyproject.toml
[project]
name = "class-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.1.0",
]
